% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewGeom.R
\name{addNewGeoms}
\alias{addNewGeoms}
\title{Add New Country Geometries}
\usage{
addNewGeoms()
}
\value{
world Map with new geoms added
}
\description{
Add New Country Geometries
}
\examples{
addNewGeoms()
}
