% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTable.R
\name{calc_risk}
\alias{calc_risk}
\title{Calculate Risk}
\usage{
calc_risk(p_I, g)
}
\arguments{
\item{p_I}{Probability one individual in a population is infectious.}

\item{g}{Event size.}
}
\value{
The risk (\%) one or more individuals at an event of size g will be infectious.
}
\description{
Calculate Risk
}
\examples{
\dontrun{
risk <- calcrisk(.001, 50)
}

}
\seealso{
\code{\link[=estRisk]{estRisk()}}
}
