% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTable.R
\name{calcRisk}
\alias{calcRisk}
\title{Calculate Risk}
\usage{
calcRisk(p_I, g)
}
\arguments{
\item{p_I}{Probability one individual in a population is infectious.}

\item{g}{Event size.}
}
\value{
The risk (\%) one or more individuals at an event of size g will be infectious.
}
\description{
Calculates the percentage probability that one or more persons in a group, of a particular size g, may be infectious given the underlying prevalence of disease.
}
\examples{
risk <- calcRisk(.001, 50)

}
\seealso{
\code{\link[=estRisk]{estRisk()}}
}
