% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTable.R
\name{create_c19r_data}
\alias{create_c19r_data}
\title{Create Table of}
\usage{
create_c19r_data(
  df_in,
  risk_output = sprintf("world_risk_regions/world_risk_regions_\%s.csv",
    str_replace_all(lubridate::today(), "-", "")),
  output_prefix = ".",
  event_size = c(10, 15, 20, 25, 50, 100, 500, 1000, 5000),
  asc_bias_list = c(3, 4, 5)
)
}
\arguments{
\item{asc_bias_list}{}
}
\value{

}
\description{
Create Table of
}
