% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadItaly.R
\name{dataQueryItaly}
\alias{dataQueryItaly}
\title{dataQueryItaly}
\usage{
dataQueryItaly(date)
}
\arguments{
\item{date}{search for Italian COVID-19 data on this date}
}
\value{
returns data for Italy on specified date. Called by LoadItaly().
}
\description{
dataQueryItaly
}
\keyword{internal}
