% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadUK.R
\name{dataQueryUK}
\alias{dataQueryUK}
\title{dataQueryUK}
\usage{
dataQueryUK(date)
}
\arguments{
\item{date}{Get COVID-19 data for this date.}
}
\value{
COVID-19 data for UK on this date. Used in LoadUK().
}
\description{
dataQueryUK
}
\keyword{internal}
