% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingFunctions.R
\name{EventMap_tmap}
\alias{EventMap_tmap}
\title{Title}
\usage{
EventMap_tmap(
  DATA,
  G,
  boundaryweights = 0.05,
  projectionCRS = "+proj=eqearth",
  maptitle = NA
)
}
\arguments{
\item{maptitle}{}
}
\value{

}
\description{
Title
}
