% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomAfghanistan}
\alias{geomAfghanistan}
\title{geomAfghanistan}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 34 rows and 9 columns.
}
\usage{
geomAfghanistan
}
\description{
geomAfghanistan
}
\keyword{datasets}
