% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomAlgeria}
\alias{geomAlgeria}
\title{geomAlgeria}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 48 rows and 9 columns.
}
\usage{
geomAlgeria
}
\description{
geomAlgeria
}
\keyword{datasets}
