% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomArgentina}
\alias{geomArgentina}
\title{geomArgentina}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 24 rows and 9 columns.
}
\usage{
geomArgentina
}
\description{
geomArgentina
}
\keyword{datasets}
