% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomAustralia}
\alias{geomAustralia}
\title{geomAustralia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 9 columns.
}
\usage{
geomAustralia
}
\description{
geomAustralia
}
\keyword{datasets}
