% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomAustria}
\alias{geomAustria}
\title{geomAustria}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 94 rows and 9 columns.
}
\usage{
geomAustria
}
\description{
geomAustria
}
\keyword{datasets}
