% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomBelgium}
\alias{geomBelgium}
\title{geomBelgium}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 43 rows and 9 columns.
}
\usage{
geomBelgium
}
\description{
geomBelgium
}
\keyword{datasets}
