% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomBrazil}
\alias{geomBrazil}
\title{geomBrazil}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 27 rows and 9 columns.
}
\usage{
geomBrazil
}
\description{
geomBrazil
}
\keyword{datasets}
