% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomCanada}
\alias{geomCanada}
\title{geomCanada}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 92 rows and 9 columns.
}
\usage{
geomCanada
}
\description{
geomCanada
}
\keyword{datasets}
