% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomChina}
\alias{geomChina}
\title{geomChina}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 33 rows and 9 columns.
}
\usage{
geomChina
}
\description{
geomChina
}
\keyword{datasets}
