% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomColombia}
\alias{geomColombia}
\title{geomColombia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 33 rows and 9 columns.
}
\usage{
geomColombia
}
\description{
geomColombia
}
\keyword{datasets}
