% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomCuba}
\alias{geomCuba}
\title{geomCuba}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 16 rows and 9 columns.
}
\usage{
geomCuba
}
\description{
geomCuba
}
\keyword{datasets}
