% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomCzechia}
\alias{geomCzechia}
\title{geomCzechia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 77 rows and 9 columns.
}
\usage{
geomCzechia
}
\description{
geomCzechia
}
\keyword{datasets}
