% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomDenmark}
\alias{geomDenmark}
\title{geomDenmark}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 99 rows and 9 columns.
}
\usage{
geomDenmark
}
\description{
geomDenmark
}
\keyword{datasets}
