% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomEcuador}
\alias{geomEcuador}
\title{geomEcuador}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 224 rows and 9 columns.
}
\usage{
geomEcuador
}
\description{
geomEcuador
}
\keyword{datasets}
