% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomEurope}
\alias{geomEurope}
\title{geomEurope}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 539 rows and 9 columns.
}
\usage{
geomEurope
}
\description{
geomEurope
}
\keyword{datasets}
