% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomFrance}
\alias{geomFrance}
\title{geomFrance}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 101 rows and 9 columns.
}
\usage{
geomFrance
}
\description{
geomFrance
}
\keyword{datasets}
