% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomGermany}
\alias{geomGermany}
\title{geomGermany}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 411 rows and 9 columns.
}
\usage{
geomGermany
}
\description{
geomGermany
}
\keyword{datasets}
