% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomGhana}
\alias{geomGhana}
\title{geomGhana}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 16 rows and 9 columns.
}
\usage{
geomGhana
}
\description{
geomGhana
}
\keyword{datasets}
