% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomGlobal}
\alias{geomGlobal}
\title{geomGlobal}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9493 rows and 10 columns.
}
\usage{
geomGlobal
}
\description{
geomGlobal
}
\keyword{datasets}
