% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomGlobal_simplified}
\alias{geomGlobal_simplified}
\title{geomGlobal_simplified}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 9477 rows and 10 columns.
}
\usage{
geomGlobal_simplified
}
\description{
geomGlobal_simplified
}
\keyword{datasets}
