% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomHaiti}
\alias{geomHaiti}
\title{geomHaiti}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 9 columns.
}
\usage{
geomHaiti
}
\description{
geomHaiti
}
\keyword{datasets}
