% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomIndia}
\alias{geomIndia}
\title{geomIndia}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36 rows and 9 columns.
}
\usage{
geomIndia
}
\description{
geomIndia
}
\keyword{datasets}
