% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomIndonesia}
\alias{geomIndonesia}
\title{geomIndonesia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 34 rows and 9 columns.
}
\usage{
geomIndonesia
}
\description{
geomIndonesia
}
\keyword{datasets}
