% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomIreland}
\alias{geomIreland}
\title{geomIreland}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 26 rows and 9 columns.
}
\usage{
geomIreland
}
\description{
geomIreland
}
\keyword{datasets}
