% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomItaly}
\alias{geomItaly}
\title{geomItaly}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 107 rows and 9 columns.
}
\usage{
geomItaly
}
\description{
geomItaly
}
\keyword{datasets}
