% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomJapan}
\alias{geomJapan}
\title{geomJapan}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 47 rows and 9 columns.
}
\usage{
geomJapan
}
\description{
geomJapan
}
\keyword{datasets}
