% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomMexico}
\alias{geomMexico}
\title{geomMexico}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2465 rows and 9 columns.
}
\usage{
geomMexico
}
\description{
geomMexico
}
\keyword{datasets}
