% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomNetherlands}
\alias{geomNetherlands}
\title{geomNetherlands}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 352 rows and 9 columns.
}
\usage{
geomNetherlands
}
\description{
geomNetherlands
}
\keyword{datasets}
