% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomNigeria}
\alias{geomNigeria}
\title{geomNigeria}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 37 rows and 9 columns.
}
\usage{
geomNigeria
}
\description{
geomNigeria
}
\keyword{datasets}
