% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomNorway}
\alias{geomNorway}
\title{geomNorway}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 356 rows and 9 columns.
}
\usage{
geomNorway
}
\description{
geomNorway
}
\keyword{datasets}
