% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomNZ}
\alias{geomNZ}
\title{geomNZ}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 20 rows and 9 columns.
}
\usage{
geomNZ
}
\description{
geomNZ
}
\keyword{datasets}
