% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomOman}
\alias{geomOman}
\title{geomOman}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 11 rows and 9 columns.
}
\usage{
geomOman
}
\description{
geomOman
}
\keyword{datasets}
