% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomPakistan}
\alias{geomPakistan}
\title{geomPakistan}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 9 columns.
}
\usage{
geomPakistan
}
\description{
geomPakistan
}
\keyword{datasets}
