% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomPeru}
\alias{geomPeru}
\title{geomPeru}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 9 columns.
}
\usage{
geomPeru
}
\description{
geomPeru
}
\keyword{datasets}
