% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomPhilippines}
\alias{geomPhilippines}
\title{geomPhilippines}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 84 rows and 9 columns.
}
\usage{
geomPhilippines
}
\description{
geomPhilippines
}
\keyword{datasets}
