% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomSaudiArabia}
\alias{geomSaudiArabia}
\title{geomSaudiArabia}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 13 rows and 9 columns.
}
\usage{
geomSaudiArabia
}
\description{
geomSaudiArabia
}
\keyword{datasets}
