% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomSmallCountries}
\alias{geomSmallCountries}
\title{geomSmallCountries}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 136 rows and 9 columns.
}
\usage{
geomSmallCountries
}
\description{
geomSmallCountries
}
\keyword{datasets}
