% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomSouthAfrica}
\alias{geomSouthAfrica}
\title{geomSouthAfrica}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 9 rows and 9 columns.
}
\usage{
geomSouthAfrica
}
\description{
geomSouthAfrica
}
\keyword{datasets}
