% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomSpain}
\alias{geomSpain}
\title{geomSpain}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 52 rows and 9 columns.
}
\usage{
geomSpain
}
\description{
geomSpain
}
\keyword{datasets}
