% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomSweden}
\alias{geomSweden}
\title{geomSweden}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 21 rows and 9 columns.
}
\usage{
geomSweden
}
\description{
geomSweden
}
\keyword{datasets}
