% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomSwitzerlandLiechtenstein}
\alias{geomSwitzerlandLiechtenstein}
\title{geomSwitzerlandLiechtenstein}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 27 rows and 9 columns.
}
\usage{
geomSwitzerlandLiechtenstein
}
\description{
geomSwitzerlandLiechtenstein
}
\keyword{datasets}
