% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomTaiwan}
\alias{geomTaiwan}
\title{geomTaiwan}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 22 rows and 9 columns.
}
\usage{
geomTaiwan
}
\description{
geomTaiwan
}
\keyword{datasets}
