% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomThailand}
\alias{geomThailand}
\title{geomThailand}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 77 rows and 9 columns.
}
\usage{
geomThailand
}
\description{
geomThailand
}
\keyword{datasets}
