% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomUnitedKingdom}
\alias{geomUnitedKingdom}
\title{geomUnitedKingdom}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 214 rows and 9 columns.
}
\usage{
geomUnitedKingdom
}
\description{
geomUnitedKingdom
}
\keyword{datasets}
