% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomUnitedStates}
\alias{geomUnitedStates}
\title{geomUnitedStates}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3229 rows and 9 columns.
}
\usage{
geomUnitedStates
}
\description{
geomUnitedStates
}
\keyword{datasets}
