% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomVenezuela}
\alias{geomVenezuela}
\title{geomVenezuela}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 9 columns.
}
\usage{
geomVenezuela
}
\description{
geomVenezuela
}
\keyword{datasets}
