% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomVietnam}
\alias{geomVietnam}
\title{geomVietnam}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 63 rows and 9 columns.
}
\usage{
geomVietnam
}
\description{
geomVietnam
}
\keyword{datasets}
