% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomZimbabwe}
\alias{geomZimbabwe}
\title{geomZimbabwe}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 9 columns.
}
\usage{
geomZimbabwe
}
\description{
geomZimbabwe
}
\keyword{datasets}
