% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadJapan.R
\name{getDataJapan}
\alias{getDataJapan}
\title{getDataJapan}
\usage{
getDataJapan(i)
}
\arguments{
\item{i}{prefecture index.}
}
\value{
COVID-19 data for Japan for this prefecture index. Used in LoadJapan().
}
\description{
getDataJapan
}
\keyword{internal}
