% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNetherlands.R
\name{getDataND}
\alias{getDataND}
\title{getDataND}
\usage{
getDataND(code)
}
\arguments{
\item{code}{Municipality code.}
}
\value{
COVID-19 data for the Netherlands. Used in LoadNetherlands().
}
\description{
getDataND
}
\keyword{internal}
