% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPhilippines.R
\name{getDataPH}
\alias{getDataPH}
\title{getDataPH}
\usage{
getDataPH(code)
}
\arguments{
\item{code}{Province Code}
}
\value{
COVID-19 data for the Philippines Used in LoadPhilippines().
}
\description{
getDataPH
}
\keyword{internal}
