% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAlgeria.R
\name{LoadAlgeria}
\alias{LoadAlgeria}
\title{LoadAlgeria}
\usage{
LoadAlgeria()
}
\value{
This function returns a spatial data frame of the current COVID-19 infection rate in Algeria
}
\description{
LoadAlgeria
}
\examples{
LoadAlgeria()
}
