% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAustria.R
\name{LoadAustria}
\alias{LoadAustria}
\title{LoadAustria}
\usage{
LoadAustria()
}
\description{
Reads in subnational data for Austria to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data is from the Federal Ministry for Social Affairs, Health, Care and Consumer Protection (BMSGPK) data on COVID-19 for Austria \url{https://www.data.gv.at/covid-19/}.
}
\examples{
\dontrun{
Austria <- LoadAustria()
}
}
\seealso{
\link{LoadData}
}
