% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBrazil.R
\name{LoadBrazil}
\alias{LoadBrazil}
\title{LoadBrazil}
\usage{
LoadBrazil()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Brazil to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
COVID-19 data are aggregated from Ministério da Saúde and Brasil.IO by \url{https://github.com/wcota/covid19br}
}
\examples{
\dontrun{
Brazil <- LoadBrazil()
}
}
\references{
W. Cota, “Monitoring the number of COVID-19 cases and deaths in brazil at municipal and federative units level”, SciELOPreprints:362 (2020), 10.1590/scielopreprints.362
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
