% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCanada.R
\name{LoadCanada}
\alias{LoadCanada}
\title{LoadCanada}
\usage{
LoadCanada()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Canada to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data aggregated by the COVID-19 Canada Open Data Working Group from local health resources at \url{https://github.com/ccodwg/Covid19Canada}.
}
\examples{
Canada <- LoadCanada()
}
\references{
Berry I, Soucy J-PR, Tuite A, Fisman D. Open access epidemiologic data and an interactive dashboard to monitor the COVID-19 outbreak in Canada. CMAJ. 2020 Apr 14;192(15):E420. doi: \url{https://doi.org/10.1503/cmaj.75262}.
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
