% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCountries.R
\name{LoadCountries}
\alias{LoadCountries}
\title{LoadCountries}
\usage{
LoadCountries(countries = NULL)
}
\arguments{
\item{countries}{List of countries to collect data for. Use NULL to attempt to download all available datasets.}
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
LoadCountries
}
\examples{
\dontrun{
AllData <- LoadCountries()
}
}
