% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCuba.R
\name{LoadCuba}
\alias{LoadCuba}
\title{LoadCuba}
\usage{
LoadCuba()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Cuba to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data aggregated from local health resources by Covid19CubaData \url{http://covid19cubadata.github.io/}.
}
\examples{
\dontrun{
Cuba <- LoadCuba()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
