% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{LoadData}
\usage{
LoadData(
  functionNames = NULL,
  custom = FALSE,
  filepath = NULL,
  interactiveMode = interactive(),
  tidy = TRUE,
  DaysOld = 30,
  minimumpercapitaactivecases = 0,
  RiskEval = NULL,
  dropNACountry = TRUE,
  dropNAall = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{functionNames}{Name(s) of function representing that for loading data - options are in countrylist. Use NULL to attempt to download all available datasets.}

\item{custom}{If TRUE, allows for use with functions not listed in object countrylist (experimental usage).}

\item{filepath}{Optionally, provide a filepath to save an error csv file to.}

\item{interactiveMode}{Set whether the session is being run interactively. If not and no googledrive oauth token is found, avoid data requiring googledrive auth token.}

\item{tidy}{If TRUE, then perform tidying according to other parameters. If FALSE, then do nothing. (passed to tidy_Data).}

\item{DaysOld}{Set any pInf data more than this days old to NA.(passed to tidy_Data).}

\item{minimumpercapitaactivecases}{Set any pInf data less than this to NA.(passed to tidy_Data).}

\item{RiskEval}{Set pInf to NA when risk is below RiskEval$minimumRisk (\%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases). (passed to tidy_Data).}

\item{dropNACountry}{If TRUE, remove rows for countries whose pInf estimates all return NA.(passed to tidy_Data).}

\item{dropNAall}{If TRUE, remove rows for any region whose pInf estimates all return NA. (passed to tidy_Data).}

\item{verbose}{If TRUE, reports on loading progress and returns warnings/errors to console.}
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
General framework for loading and tidying data.
}
\examples{
LoadData("LoadUS")
LoadData("LoadUS", dropNAall = TRUE)
LoadData("LoadNewZealand",tidy = FALSE)
LoadData(c("LoadUS","LoadMalaysia"))
\dontrun{
LoadData()
}
}
\seealso{
\code{\link[=tidy_Data]{tidy_Data()}}
}
