% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDenmark.R
\name{LoadDenmark}
\alias{LoadDenmark}
\title{LoadDenmark}
\usage{
LoadDenmark()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Denmark to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
COVID-19 data from the Statens Serum Institut (SSI):
\itemize{
\item \url{https://covid19.ssi.dk/overvagningsdata}
\item \url{https://experience.arcgis.com/experience/aa41b29149f24e20a4007a0c4e13db1d}.
}
}
\examples{
\dontrun{
Denmark <- LoadDenmark()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
