% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadEcuador.R
\name{LoadEcuador}
\alias{LoadEcuador}
\title{LoadEcuador}
\usage{
LoadEcuador()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Ecuador to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data compiled from the Servicio Nacional de Gestión de Riesgos y Emergencias del Ecuador by Ecuacovid: \url{https://github.com/andrab/ecuacovid}.
}
\examples{
Ecuador <- LoadEcuador()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
