% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadErrorWarning.R
\name{LoadErrorWarning}
\alias{LoadErrorWarning}
\title{LoadErrorWarning}
\usage{
LoadErrorWarning(text)
}
\arguments{
\item{text}{Text to use in warning message - here, should be of form LoadX, where X is the data that is attempted to be loaded.}
}
\value{
prints warning.
}
\description{
Displays warning message to note that loading failed. To be used internally within LoadX type function calls.
}
\examples{
LoadErrorWarning("testing_LoadErrorWarning_message")
}
\keyword{internal}
