% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGermany.R
\name{LoadGermany}
\alias{LoadGermany}
\title{LoadGermany}
\usage{
LoadGermany()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Germany to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
COVID-19 case data for Germany is from the Robert Koch-Institut and the Bundesamt für Kartographie und Geodäsie: \url{https://npgeo-corona-npgeo-de.hub.arcgis.com/} \url{https://npgeo-corona-npgeo-de.hub.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0/about}.
}
\examples{
\dontrun{
Germany <- LoadGermany()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
