% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndia.R
\name{LoadIndia}
\alias{LoadIndia}
\title{LoadIndia}
\usage{
LoadIndia()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for India to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data collated by \url{https://covid19tracker.in/}, an initiative of the Indian Institute of Technology Hyderabad, from state bulletins and official reports with thanks to the covid19india.org team for their outstanding work in creating the original portal, and for making their code base public.
}
\examples{
\dontrun{
India <- LoadIndia()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
