% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndonesia.R
\name{LoadIndonesia}
\alias{LoadIndonesia}
\title{LoadIndonesia}
\usage{
LoadIndonesia()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Indonesia to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Live data, interactive charts & maps of Indonesia provincial COVID-19 daily cases and vaccination. Live version: \url{https://erlange.github.io/INACOVID/}
\url{https://github.com/erlange/INACOVID}.
}
\examples{
Indonesia <- LoadIndonesia()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
