% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIreland.R
\name{LoadIreland}
\alias{LoadIreland}
\title{LoadIreland}
\usage{
LoadIreland()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Ireland to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data is provided by the Health Service Executive (HSE), Health Protection Surveillance Centre (HPSC), The Central Statistics Office (CSO) and Gov.ie and accessed via Ireland's COVID-19 Data Hub: \url{https://covid19ireland-geohive.hub.arcgis.com/}.
}
\examples{
Ireland <- LoadIreland()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
