% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadItaly.R
\name{LoadItaly}
\alias{LoadItaly}
\title{LoadItaly}
\usage{
LoadItaly()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Italy to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data collated by the Italian Department of Civil Protection COVID-19: \url{https://github.com/pcm-dpc/COVID-19/}.
}
\examples{
Italy <- LoadItaly()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
