% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadJHUCSSE.R
\name{LoadJHUCSSE}
\alias{LoadJHUCSSE}
\title{LoadJHUCSSE}
\usage{
LoadJHUCSSE()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for various regions in the JHU CSSE COVID-19 Dataset to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data Aggregated by the COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University \url{https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data}.
}
\examples{
JH <- LoadJHUCSSE()
}
\references{
Dong, E., Du, H., & Gardner, L. (2020). An interactive web-based dashboard to track COVID-19 in real time. The Lancet infectious diseases, 20(5), 533-534.
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
