% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMexico.R
\name{LoadMexico}
\alias{LoadMexico}
\title{LoadMexico}
\usage{
LoadMexico()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Mexico to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
COVID-19 data is sourced from the Covid-19 México hub page: \url{https://datos.covid-19.conacyt.mx/}.
}
\examples{
\dontrun{
Mexico <- LoadMexico()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
