% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNigeria.R
\name{LoadNigeria}
\alias{LoadNigeria}
\title{LoadNigeria}
\usage{
LoadNigeria()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Nigeria to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
COVID-19 data collated by Humanitarian Emergency Response Africa  \url{https://data.humdata.org/dataset/nigeria_covid19_subnational} which is
sourced from data collected by Nigeria Centre for Disease Control and National Primary Health Care Development Agency.
}
\examples{
Nigeria <- LoadNigeria()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
