% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNorway.R
\name{LoadNorway}
\alias{LoadNorway}
\title{LoadNorway}
\usage{
LoadNorway()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Norway to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Thomas, Haarstad, F., Manuel & YBK. Public COVID-19 Data for Norway (covid19data.no). \url{https://github.com/thohan88/covid19-nor-data}.
}
\examples{
Norway <- LoadNorway()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
