% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadOmanTESTING.R
\name{LoadOmanTESTING}
\alias{LoadOmanTESTING}
\title{LoadOmanTESTING}
\usage{
LoadOmanTESTING(
  tidy = TRUE,
  DaysOld = 30,
  minimumpercapitaactivecases = 0,
  RiskEval = NULL,
  dropNACountry = FALSE,
  dropNAall = FALSE
)
}
\arguments{
\item{tidy}{If TRUE, then perform tidying according to other parameters. If FALSE, then do nothing.}

\item{DaysOld}{Set any pInf data more than this days old to NA.}

\item{minimumpercapitaactivecases}{Set any pInf data less than this to NA.}

\item{RiskEval}{Set pInf to NA when risk is below RiskEval$minimumRisk (\%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases).}

\item{dropNACountry}{If TRUE, remove rows for countries whose pInf estimates all return NA.}

\item{dropNAall}{If TRUE, remove rows for any region whose pInf estimates all return NA}
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Oman to calculate most recent estimate of per capita active COVID-19 cases.
}
\note{
Ministry of Health for Oman, collated by Safeture for the Humanitarian Data Exchange: \url{https://data.humdata.org/dataset/oman-coronavirus-covid-19-subnational-cases}.
}
\examples{
Oman2 <- LoadOmanTESTING()
}
\seealso{
\code{\link[=LoadCountries]{LoadCountries()}}
}
