% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPakistan.R
\name{LoadPakistan}
\alias{LoadPakistan}
\title{LoadPakistan}
\usage{
LoadPakistan()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Pakistan to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data obtained from COVID-19 Data Hub \url{https://covid19datahub.io}
}
\examples{
Pakistan <- LoadPakistan()
}
\references{
Guidotti et al., (2020). COVID-19 Data Hub. Journal of Open Source Software, 5(51), 2376, \url{https://doi.org/10.21105/joss.02376}.
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
