% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPhilippines.R
\name{LoadPhilippines}
\alias{LoadPhilippines}
\title{LoadPhilippines}
\usage{
LoadPhilippines()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Philippines to calculate most recent estimate of per capita active COVID-19 cases. Note that this process requires the use of the googledrive package which requires the users google credentials to be provided. Use with LoadData() is recommended.
}
\note{
Data obtained from the Republic of Philippines Department of Health: \url{https://doh.gov.ph/covid19tracker}.
}
\examples{
\dontrun{
Philippines <- LoadPhilippines()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
