% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSaudiArabia.R
\name{LoadSaudiArabia}
\alias{LoadSaudiArabia}
\title{LoadSaudiArabia}
\usage{
LoadSaudiArabia()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Saudi Arabia to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data sourced from Ministry of Health, Covid19 Command and Control Center CCC, The National Health Emergency Operation Center NHEOC; and assembled by National Health Command and Control NHCC, Covid19 Data and Informatics Committee. \url{https://covid19.moh.gov.sa/}.
}
\examples{
\dontrun{
SaudiArabia <- LoadSaudiArabia()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
