% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSouthAfrica.R
\name{LoadSouthAfrica}
\alias{LoadSouthAfrica}
\title{LoadSouthAfrica}
\usage{
LoadSouthAfrica()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for South Africa to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data Science for Social Impact Research Group @ University of Pretoria, Coronavirus COVID-19 (2019-nCoV) Data Repository for South Africa. Available on: \url{https://github.com/dsfsi/covid19za}.
}
\examples{
SouthAfrica <- LoadSouthAfrica()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
