% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSpain.R
\name{LoadSpain}
\alias{LoadSpain}
\title{LoadSpain}
\usage{
LoadSpain()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Spain to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
COVID-19 data from España Ministerio de Sanidad and Instituto de Salud Carlos III: \url{https://cnecovid.isciii.es/covid19/}.
}
\examples{
Spain <- LoadSpain()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
