% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadTaiwan.R
\name{LoadTaiwan}
\alias{LoadTaiwan}
\title{LoadTaiwan}
\usage{
LoadTaiwan()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Taiwan to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data sourced from Taiwan Centers for Disease Control \url{https://data.cdc.gov.tw/en/dataset/aagsdctable-day-19cov}.
}
\examples{
\dontrun{
Taiwan <- LoadTaiwan()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
