% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadUK.R
\name{LoadUK}
\alias{LoadUK}
\title{LoadUK}
\usage{
LoadUK()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for the United Kingdom to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
The COVID-19 data is from the UK API from Public Health England and NHSX: \url{https://coronavirus.data.gov.uk}.
}
\examples{
\dontrun{
UK <- LoadUK()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
