% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadUS.R
\name{LoadUS}
\alias{LoadUS}
\title{LoadUS}
\usage{
LoadUS()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for the United States to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Real-time county level COVID19 data comes from the NYTimes COVID19 data project: \url{https://github.com/nytimes/covid-19-data}.
US 2019 population estimate data comes from the US Census: \url{https://www.census.gov/data/tables/time-series/demo/popest/2010s-state-total.html}.
}
\examples{
US <- LoadUS()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
