% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadZimbabwe.R
\name{LoadZimbabwe}
\alias{LoadZimbabwe}
\title{LoadZimbabwe}
\usage{
LoadZimbabwe()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Zimbabwe to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Sourced from the COVID-19 Data Repository by African Surveyors Connect \url{https://github.com/African-Surveyors-Connect/Zimbabwe-COVID-19-Data/}.
}
\examples{
Zimbabwe <- LoadZimbabwe()
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
