% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m49}
\alias{m49}
\title{m49}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 249 rows and 16 columns.
}
\usage{
m49
}
\description{
m49
}
\keyword{datasets}
