% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_argentina}
\alias{misc_argentina}
\title{misc_argentina}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24 rows and 10 columns.
}
\usage{
misc_argentina
}
\description{
misc_argentina
}
\keyword{datasets}
