% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_canada}
\alias{misc_canada}
\title{misc_canada}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 92 rows and 37 columns.
}
\usage{
misc_canada
}
\description{
misc_canada
}
\keyword{datasets}
