% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_chile}
\alias{misc_chile}
\title{misc_chile}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 8 columns.
}
\usage{
misc_chile
}
\description{
misc_chile
}
\keyword{datasets}
