% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_china}
\alias{misc_china}
\title{misc_china}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 9 columns.
}
\usage{
misc_china
}
\description{
misc_china
}
\keyword{datasets}
