% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_ireland}
\alias{misc_ireland}
\title{misc_ireland}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26 rows and 15 columns.
}
\usage{
misc_ireland
}
\description{
misc_ireland
}
\keyword{datasets}
