% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_netherlands}
\alias{misc_netherlands}
\title{misc_netherlands}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 352 rows and 24 columns.
}
\usage{
misc_netherlands
}
\description{
misc_netherlands
}
\keyword{datasets}
