% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_saudiarabia}
\alias{misc_saudiarabia}
\title{misc_saudiarabia}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13 rows and 13 columns.
}
\usage{
misc_saudiarabia
}
\description{
misc_saudiarabia
}
\keyword{datasets}
