% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_southkorea}
\alias{misc_southkorea}
\title{misc_southkorea}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 13 columns.
}
\usage{
misc_southkorea
}
\description{
misc_southkorea
}
\keyword{datasets}
