% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_thailand}
\alias{misc_thailand}
\title{misc_thailand}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 77 rows and 7 columns.
}
\usage{
misc_thailand
}
\description{
misc_thailand
}
\keyword{datasets}
