% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_uk}
\alias{misc_uk}
\title{misc_uk}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 214 rows and 10 columns.
}
\usage{
misc_uk
}
\description{
misc_uk
}
\keyword{datasets}
