% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_vietnam}
\alias{misc_vietnam}
\title{misc_vietnam}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 63 rows and 4 columns.
}
\usage{
misc_vietnam
}
\description{
misc_vietnam
}
\keyword{datasets}
