% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miscPakistan}
\alias{miscPakistan}
\title{miscPakistan}
\format{
An object of class \code{data.frame} with 7 rows and 8 columns.
}
\usage{
miscPakistan
}
\description{
miscPakistan
}
\keyword{datasets}
