% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewGeom.R
\name{moveFiles}
\alias{moveFiles}
\title{Move Processed Files}
\usage{
moveFiles(x)
}
\arguments{
\item{x}{File name to be moved}
}
\description{
Move Processed Files
}
\examples{
file.create(here::here("tools", "toProcess", "testFile.R"))
moveFiles("testFile.R")
}
\keyword{internal}
