% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_belgium}
\alias{names_belgium}
\title{names_belgium}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 44 rows and 2 columns.
}
\usage{
names_belgium
}
\description{
names_belgium
}
\keyword{datasets}
