% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_europe}
\alias{names_europe}
\title{names_europe}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 755 rows and 4 columns.
}
\usage{
names_europe
}
\description{
names_europe
}
\keyword{datasets}
