% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_philippines}
\alias{names_philippines}
\title{names_philippines}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1647 rows and 11 columns.
}
\usage{
names_philippines
}
\description{
names_philippines
}
\keyword{datasets}
