% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_taiwan}
\alias{names_taiwan}
\title{names_taiwan}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 22 rows and 4 columns.
}
\usage{
names_taiwan
}
\description{
names_taiwan
}
\keyword{datasets}
