% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingFunctions.R
\name{PerCapitaMap_leaflet}
\alias{PerCapitaMap_leaflet}
\title{Title}
\usage{
PerCapitaMap_leaflet(DATA, people, boundaryweights = 0.05)
}
\arguments{
\item{boundaryweights}{}
}
\value{

}
\description{
Title
}
