% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_algeria}
\alias{pop_algeria}
\title{pop_algeria}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 48 rows and 2 columns.
}
\usage{
pop_algeria
}
\description{
pop_algeria
}
\keyword{datasets}
