% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_brazil}
\alias{pop_brazil}
\title{pop_brazil}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5570 rows and 11 columns.
}
\usage{
pop_brazil
}
\description{
pop_brazil
}
\keyword{datasets}
