% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_canada}
\alias{pop_canada}
\title{pop_canada}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 102 rows and 7 columns.
}
\usage{
pop_canada
}
\description{
pop_canada
}
\keyword{datasets}
