% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_saudiarabia}
\alias{pop_saudiarabia}
\title{pop_saudiarabia}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13 rows and 5 columns.
}
\usage{
pop_saudiarabia
}
\description{
pop_saudiarabia
}
\keyword{datasets}
