% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_smallcountries}
\alias{pop_smallcountries}
\title{pop_smallcountries}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 251 rows and 53 columns.
}
\usage{
pop_smallcountries
}
\description{
pop_smallcountries
}
\keyword{datasets}
