% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_southafrica}
\alias{pop_southafrica}
\title{pop_southafrica}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
pop_southafrica
}
\description{
pop_southafrica
}
\keyword{datasets}
