% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_venezuela}
\alias{pop_venezuela}
\title{pop_venezuela}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25 rows and 3 columns.
}
\usage{
pop_venezuela
}
\description{
pop_venezuela
}
\keyword{datasets}
