% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_zimbabwe}
\alias{pop_zimbabwe}
\title{pop_zimbabwe}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 8 columns.
}
\usage{
pop_zimbabwe
}
\description{
pop_zimbabwe
}
\keyword{datasets}
