% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomUpdateUtils.R
\name{remGeoSurplus}
\alias{remGeoSurplus}
\title{Remove Surplus Geometries}
\usage{
remGeoSurplus(
  input_file,
  collection_files = c("geomEurope", "geomSmallCountries")
)
}
\arguments{
\item{input_file}{In file}

\item{collection_files}{File(s) with geoms to remove}
}
\value{
input_file Updated input file
}
\description{
Some geography files are collections of many countries that overlap with other files that have their own higher-quality data. The remGeoSurplus function ensures each country is only present once in the dataset.
}
\examples{
\dontrun{
data("geomGlobal")
localcovid19now:::remGeoSurplus(geomGlobal)
}
}
\keyword{internal}
