% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remSurplus.R
\name{remSurplus}
\alias{remSurplus}
\title{Remove Surplus Geometries}
\usage{
remSurplus(
  input_file,
  collection_files = c("geomEurope", "geomSmallCountries")
)
}
\arguments{
\item{input_file}{}

\item{collection_files}{}
}
\value{

}
\description{
Remove Surplus Geometries
}
\examples{
data("geomGlobal")
remSurplus(geomGlobal)
}
