% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_Data.R
\name{tidy_Data}
\alias{tidy_Data}
\title{tidy_Data}
\usage{
tidy_Data(
  DATA,
  DaysOld = 30,
  minimumpercapitaactivecases = 0,
  RiskEval = NULL,
  dropNACountry = FALSE,
  dropNAall = FALSE
)
}
\arguments{
\item{DATA}{Data created using a LoadCountry call.}
}
\value{
Returns input simple feature with NA values for pInf when reports are over 30 days old or suggest that there is no, or negative, change in cumulative cases over the most recent 14 day period.
}
\description{
For an object read in with a LoadCountry function, sets the per capita active cases (pInf) to NA when reports are over 30 days old, or if no, or negative, change in cumulative cases over the most recent 14 day period.
}
\examples{
\dontrun{
Philippines <- LoadPhilippines()
Philippines <- tidy_Data(Philippines)
}
}
