% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomUpdateUtils.R
\name{updateGlobal}
\alias{updateGlobal}
\title{Update Global Geometries}
\usage{
updateGlobal()
}
\description{
Function to integrate in new geometries and update the geometries across the package
}
\examples{
\dontrun{
localcovid19now:::updateGlobal()
}
}
\keyword{internal}
