% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingFunctions.R
\name{EventMap_leaflet}
\alias{EventMap_leaflet}
\title{EventMap_leaflet}
\usage{
EventMap_leaflet(DATA, G, AB, boundaryweights = 0.05)
}
\arguments{
\item{DATA}{Data containing prevalence information to map.}

\item{G}{Event size to compute risk for.}

\item{AB}{Case ascertainment bias to compute risk for.}

\item{boundaryweights}{Weight assigned to the maps boundary edges.}
}
\value{
Outputs an interactive leaflet map displaying exposure risk for the input data.
}
\description{
Creates an interactive Leaflet map displaying exposure risk for the input data.
}
\examples{
\dontrun{
Austria <- LoadData("LoadAustria")
Austria$AB <- 4
EventMap_leaflet(Austria, 50, Austria$AB)
}
}
\seealso{
Other mapplots: 
\code{\link{EventMap_tmap}()},
\code{\link{PerCapitaMap_leaflet}()},
\code{\link{PerCapitaMap_tmap}()}
}
\concept{mapplots}
