% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingFunctions.R
\name{EventMap_tmap}
\alias{EventMap_tmap}
\title{EventMap_tmap}
\usage{
EventMap_tmap(
  DATA,
  G,
  AB,
  boundaryweights = 0.05,
  projectionCRS = "+proj=eqearth",
  maptitle = NA
)
}
\arguments{
\item{DATA}{Data containing prevalence information to map.}

\item{G}{Event size to compute risk for.}

\item{AB}{Case ascertainment bias to compute risk for.}

\item{boundaryweights}{Weight assigned to the maps boundary edges.}

\item{projectionCRS}{Type of geographic projection to use.}

\item{maptitle}{Adds title to map.}
}
\value{
Outputs a tmap displaying exposure risk for the input data.
}
\description{
Creates a tmap map displaying active cases per 'people' for the input data.
}
\examples{
\dontrun{
Austria <- LoadData("LoadAustria")
Austria$AB <- 4
EventMap_tmap(Austria, 50, Austria$AB)
}
}
\seealso{
Other mapplots: 
\code{\link{EventMap_leaflet}()},
\code{\link{PerCapitaMap_leaflet}()},
\code{\link{PerCapitaMap_tmap}()}
}
\concept{mapplots}
