% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAlgeria.R
\name{LoadAlgeria}
\alias{LoadAlgeria}
\title{LoadAlgeria}
\usage{
LoadAlgeria()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Algeria to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Data was obtained using the Algeria Coronavirus Tracker API \url{https://corona-dz.live/}, documented at \url{https://github.com/Amine27/covid-19-dz} whcih sources information from the Algerian Ministry of Health, Population and Hospital Reform website.
}
\examples{
\dontrun{
Algeria <- LoadAlgeria()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
