% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNetherlands.R
\name{LoadNetherlands}
\alias{LoadNetherlands}
\title{LoadNetherlands}
\usage{
LoadNetherlands()
}
\value{
A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
}
\description{
Reads in subnational data for Netherlands to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
}
\note{
Covid-19 numbers per municipality as of publication date. RIVM / I & V / EPI. OSIRIS General Infectious Diseases (AIZ). \url{https://data.rivm.nl/geonetwork/srv/dut/catalog.search#/metadata/5f6bc429-1596-490e-8618-1ed8fd768427?tab=general}.
}
\examples{
\dontrun{
Netherlands <- LoadNetherlands()
}
}
\seealso{
\code{\link[=LoadData]{LoadData()}}
}
