% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingFunctions.R
\name{PerCapitaMap_tmap}
\alias{PerCapitaMap_tmap}
\title{PerCapitaMap_leaflet}
\usage{
PerCapitaMap_tmap(
  DATA,
  people = 1e+05,
  boundaryweights = 0.05,
  projectionCRS = "+proj=eqearth",
  maptitle = NA
)
}
\arguments{
\item{DATA}{Data containing prevalence information to map.}

\item{people}{Transform from proportion of population to per 'people'.}

\item{boundaryweights}{Weight assigned to the maps boundary edges.}

\item{projectionCRS}{Type of geographic projection to use.}

\item{maptitle}{Adds title to map.}
}
\value{
Outputs a tmap displaying active cases per 'people' for the input data.
}
\description{
PerCapitaMap_leaflet
}
\seealso{
Other mapplots: 
\code{\link{EventMap_leaflet}()},
\code{\link{EventMap_tmap}()},
\code{\link{PerCapitaMap_leaflet}()}
}
\concept{mapplots}
