% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remSurplus.R
\name{remSurplus}
\alias{remSurplus}
\title{Remove Surplus Geometries}
\usage{
remSurplus(
  input_file,
  collection_files = c("geomEurope", "geomSmallCountries")
)
}
\arguments{
\item{input_file}{In file}

\item{collection_files}{File(s) with geoms to remove}
}
\value{
input_file Updated input file
}
\description{
Remove Surplus Geometries
}
\examples{
\dontrun{
data("geomGlobal")
remSurplus(geomGlobal)
}
}
