% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_Data.R
\name{tidy_Data}
\alias{tidy_Data}
\title{tidy_Data}
\usage{
tidy_Data(
  DATA,
  DaysOld = 30,
  minimumpercapitaactivecases = 0,
  RiskEval = NULL,
  dropNACountry = FALSE,
  dropNAall = FALSE
)
}
\arguments{
\item{DATA}{Data created using a LoadCountry call.}

\item{DaysOld}{Set any pInf data more than this days old to NA.}

\item{minimumpercapitaactivecases}{Set any pInf data less than this to NA.}

\item{RiskEval}{Set pInf to NA when risk is below RiskEval$minimumRisk (\%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases).}

\item{dropNACountry}{If TRUE, remove rows for countries whose pInf estimates all return NA.}

\item{dropNAall}{If TRUE, remove rows for any region whose pInf estimates all return NA}
}
\value{
Returns input simple feature with NA values for pInf according to input options.
}
\description{
For an object read in with a LoadCountry function, sets the per capita active cases (pInf) to NA or removes them from the dataset based on input conditions. Can be used to remove data that is not recent, or data that produces low or negative estimates of per capita cases, or risk. Can also be used to remove data that is NA by country or by region.
}
\examples{
\dontrun{
Philippines <- LoadPhilippines()
Philippines <- tidy_Data(Philippines)
}
}
